+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/types/entries/ResourcefulConfigObjectEntry.java
@@ -1,5 +1,7 @@
 package com.teamresourceful.resourcefulconfig.api.types.entries;
 
+import net.minecraft.network.chat.Component;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.LinkedHashMap;
@@ -9,8 +11,17 @@ public interface ResourcefulConfigObjectEntry extends ResourcefulConfigEntry {
     @NotNull
     LinkedHashMap<String, ResourcefulConfigEntry> entries();
 
+    /**
+     * @deprecated This method was only ever intended for grabbing information like title of the object nothing else, this was an oversight to include it.
+     */
+    @Deprecated
+    @ApiStatus.ScheduledForRemoval(inVersion = "22")
     default Object instance() {
         return null;
+    }
+
+    default Component getTitle(@NotNull Component fallback) {
+        return fallback;
     }
 
 }