+++ changelog.md
@@ -1,3 +1,3 @@
-- Added a line under separators
-- Made categories actually hide when @Hidden was added to them
-- Made configs with @Hidden not show up in global list
+- stack overflow in addons
+- deprecate instance value in object entries (this was just used for the title and should have not been exposed)
+- add new title method for objects